refactor: move reviews lib
