+++ starters/shopify-meilisearch/app/actions/reviews.actions.ts
@@ -1,6 +1,6 @@
 "use server"
 
-import { reviewsClient } from "clients/reviews"
+import { createProductReview } from "lib/reviews"
 import type { ProductReviewBody } from "lib/reviews/types"
 
 import { headers } from "next/headers"
@@ -8,7 +8,7 @@ import { headers } from "next/headers"
 export const submitReview = async (payload: Omit<ProductReviewBody, "ip_addr">) => {
   try {
     const ipAddress = headers().get("x-forwarded-for") || null
-    await reviewsClient.createProductReview({ ...payload, ip_addr: ipAddress })
+    await createProductReview({ ...payload, ip_addr: ipAddress })
   } catch (err) {
     throw new Error(err as string)
   }
